feat: Rename project from VneTemplate to VneTestbed, update configuration files, documentation, and examples accordingly. Adjust versioning in release manifest and changelog. Introduce new testbed-specific examples and update build scripts for consistency.
