+++ .github/workflows/ci.yml
@@ -87,7 +87,7 @@ jobs:
           -DCMAKE_BUILD_TYPE=Release \
           -DCMAKE_C_COMPILER=${{ matrix.c_compiler }} \
           -DCMAKE_CXX_COMPILER=${{ matrix.cpp_compiler }} \
-          -DVNE_TEMPLATE_CI=ON
+          -DVNE_TESTBED_CI=ON
 
     - name: Build
       run: cmake --build build --config Release --parallel
@@ -114,7 +114,7 @@ jobs:
           -DCMAKE_BUILD_TYPE=Debug \
           -DCMAKE_C_COMPILER=gcc \
           -DCMAKE_CXX_COMPILER=g++ \
-          -DVNE_TEMPLATE_CI=ON \
+          -DVNE_TESTBED_CI=ON \
           -DWARNINGS_AS_ERRORS=ON \
           -DENABLE_COVERAGE=ON
 
@@ -179,7 +179,7 @@ jobs:
       run: |
         cmake -B build -S . \
           -DCMAKE_BUILD_TYPE=Release \
-          -DVNE_TEMPLATE_CI=ON \
+          -DVNE_TESTBED_CI=ON \
           -DWARNINGS_AS_ERRORS=ON
 
     - name: Build
@@ -222,7 +222,7 @@ jobs:
       run: |
         cmake -B build -S . `
           -DCMAKE_BUILD_TYPE=Release `
-          -DVNE_TEMPLATE_CI=ON `
+          -DVNE_TESTBED_CI=ON `
           -DWARNINGS_AS_ERRORS=ON
 
     - name: Build
@@ -263,7 +263,7 @@ jobs:
           -DCMAKE_BUILD_TYPE=Release \
           -DCMAKE_C_COMPILER=clang \
           -DCMAKE_CXX_COMPILER=clang++ \
-          -DVNE_TEMPLATE_CI=ON \
+          -DVNE_TESTBED_CI=ON \
           -DWARNINGS_AS_ERRORS=ON
 
     - name: Build
