+++ .github/workflows/codeql.yml
@@ -39,7 +39,7 @@ jobs:
       run: |
         sudo apt-get update
         sudo apt-get install -y build-essential cmake
-        cmake -B build -S . -DCMAKE_BUILD_TYPE=Release -DVNE_TEMPLATE_CI=ON -DWARNINGS_AS_ERRORS=OFF
+        cmake -B build -S . -DCMAKE_BUILD_TYPE=Release -DVNE_TESTBED_CI=ON -DWARNINGS_AS_ERRORS=OFF
         cmake --build build --config Release --parallel
 
     - name: Perform CodeQL Analysis
