+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-# Pre-commit hooks for VneTemplate
+# Pre-commit hooks for VneTestbed
 # Install: pip install pre-commit && pre-commit install
 # Run manually: pre-commit run --all-files
