+++ .release-please-manifest.json
@@ -1,3 +1,1 @@
-{
-  ".": "0.1.1"
-}
+{".":"0.0.0"}
