+++ CHANGELOG.md
@@ -7,13 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 **Note:** Entries under `[Unreleased]` are generated by [release-please](https://github.com/googleapis/release-please) from conventional commits (e.g. `feat:`, `fix:`, `docs:`). When a release PR is merged, they are moved into a versioned section and the `VERSION` file is updated.
 
-## [0.1.1](https://github.com/vertexnova/vnetemplate/compare/v0.1.0...v0.1.1) (2026-02-20)
-
-
-### Bug Fixes
-
-* **vertexnova:** Update CONTRIBUTING.md and README.md to clarify Conventional Commits usage for release management. Enhanced guidelines for commit message formatting and added instructions for initial release setup. ([#4](https://github.com/vertexnova/vnetemplate/issues/4)) ([10c5b24](https://github.com/vertexnova/vnetemplate/commit/10c5b24e9fc420d36dea721aaa56f280c493432d))
-
 ## [Unreleased]
 
 ### Added
