+++ deps/internal/README.md
@@ -8,6 +8,6 @@ VertexNova internal libraries go here. CMake modules (vnecmake) are in **cmake/v
 - **vnelogging** – Logging library (e.g. spdlog-based).  
   Clone or add as submodule into `deps/internal/vnelogging`.
 
-If these directories are missing or empty, the root CMake will skip them and the template library will build without linking to vne::common or vne::logging.
+If these directories are missing or empty, the root CMake will skip them and the testbed library will build without linking to vne::common or vne::logging.
 
 See each subdirectory’s README (once the repo is present) for recommended branch/tag and build options.
