--- docs/vertexnova/template/template.md
@@ -1,71 +1,0 @@
-# VertexNova Template
-
-## Overview
-
-VneTemplate is a minimal C++ project template for the VertexNova ecosystem. It provides a standard layout (include, src, tests, examples), CMake setup with vnecmake, optional internal deps (vnecommon, vnelogging), and a tiny API so you can build and run out of the box. Use it as a starting point for new libraries or applications.
-
-![System Context](diagrams/context.png)
-
-**Figure 1: Context Diagram**
-
-| Element | Description |
-|---------|-------------|
-| C++ Application | Developer/user code (tests, examples, or your app) that uses the template API |
-| VneTemplate | Template library; provides `get_version()`, `hello()`, and the project scaffold |
-
-## Project layout and build
-
-The template follows a standard directory layout and builds a static library, tests, and optional examples:
-
-![Project layout](diagrams/architecture.png)
-
-**Figure 2: Project layout and build**
-
-| Element | Description |
-|---------|-------------|
-| include/vertexnova/template/ | Public API headers (e.g. `template.h`) |
-| src/vertexnova/template/ | Implementation |
-| tests/ | Unit tests (Google Test) |
-| examples/ | Example apps (e.g. `01_hello_template`) |
-| cmake/vnecmake/ | CMake modules submodule |
-| deps/internal/, deps/external/ | Internal (vnecommon, vnelogging) and external (googletest) deps |
-| CMake configure + build | Produces `libvnetemplate.a`, tests, and examples |
-
-See the root [README.md](../../../README.md) for prerequisites, dependencies, and build commands.
-
-## API usage
-
-The public API lives in namespace `vne::template_ns` and exposes two functions:
-
-![API flow](diagrams/api.png)
-
-**Figure 3: API usage**
-
-| Step | Function | Description |
-|------|----------|-------------|
-| 1 | `get_version()` | Returns the project version string (e.g. from `VERSION` file). |
-| 2 | `hello()` | Returns a greeting string (minimal placeholder). |
-
-Example:
-
-```cpp
-#include <vertexnova/template/template.h>
-
-const char* ver = vne::template_ns::get_version();  // e.g. "1.0.0"
-const char* msg = vne::template_ns::hello();        // e.g. "Hello from VneTemplate"
-```
-
-## CMake options
-
-| Option | Default | Description |
-|--------|---------|-------------|
-| `VNE_TEMPLATE_TESTS` | ON | Build unit tests. |
-| `VNE_TEMPLATE_EXAMPLES` | ON (dev/top-level) / OFF (submodule) | Build examples (on by default in dev builds, off when used as a submodule). |
-| `WARNINGS_AS_ERRORS` | OFF | Treat compiler warnings as errors. |
-| `ENABLE_DOXYGEN` | OFF | Generate Doxygen documentation. |
-
-## Documentation
-
-- **This document:** `docs/vertexnova/template/template.md`
-- **Diagrams:** `docs/vertexnova/template/diagrams/` (Draw.io sources; export to PNG as described in [diagrams/README.md](diagrams/README.md))
-- **API reference:** Generated by Doxygen when `-DENABLE_DOXYGEN=ON` (see root README)
