+++ docs/vertexnova/testbed/diagrams/api.drawio
@@ -7,13 +7,13 @@
                 <mxCell id="step1" value="1. get_version()" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#e3f2fd;strokeColor=#1565c0;strokeWidth=2;fontSize=14;fontStyle=1;fontColor=#1a1a1a;" vertex="1" parent="1">
                     <mxGeometry x="80" y="80" width="140" height="36" as="geometry"/>
                 </mxCell>
-                <mxCell id="code1" value="const char* ver = vne::template_ns::get_version();&#10;// e.g. &quot;1.0.0&quot;" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#ffffff;strokeColor=#1565c0;align=left;verticalAlign=middle;spacingLeft=10;fontSize=11;fontFamily=Courier New;fontColor=#1a1a1a;" vertex="1" parent="1">
+                <mxCell id="code1" value="const char* ver = vne::testbed_ns::get_version();&#10;// e.g. &quot;1.0.0&quot;" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#ffffff;strokeColor=#1565c0;align=left;verticalAlign=middle;spacingLeft=10;fontSize=11;fontFamily=Courier New;fontColor=#1a1a1a;" vertex="1" parent="1">
                     <mxGeometry x="80" y="130" width="280" height="50" as="geometry"/>
                 </mxCell>
                 <mxCell id="step2" value="2. hello()" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#fff3e0;strokeColor=#e65100;strokeWidth=2;fontSize=14;fontStyle=1;fontColor=#1a1a1a;" vertex="1" parent="1">
                     <mxGeometry x="480" y="80" width="120" height="36" as="geometry"/>
                 </mxCell>
-                <mxCell id="code2" value="const char* msg = vne::template_ns::hello();&#10;// e.g. &quot;Hello from VneTemplate&quot;" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#ffffff;strokeColor=#e65100;align=left;verticalAlign=middle;spacingLeft=10;fontSize=11;fontFamily=Courier New;fontColor=#1a1a1a;" vertex="1" parent="1">
+                <mxCell id="code2" value="const char* msg = vne::testbed_ns::hello();&#10;// e.g. &quot;Hello from VneTestbed&quot;" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#ffffff;strokeColor=#e65100;align=left;verticalAlign=middle;spacingLeft=10;fontSize=11;fontFamily=Courier New;fontColor=#1a1a1a;" vertex="1" parent="1">
                     <mxGeometry x="480" y="130" width="320" height="50" as="geometry"/>
                 </mxCell>
                 <mxCell id="arr1" style="endArrow=blockThin;endFill=1;html=1;strokeWidth=2;strokeColor=#1565c0;exitX=1;exitY=0.5;exitDx=0;exitDy=0;entryX=0;entryY=0.5;entryDx=0;entryDy=0;" edge="1" parent="1" source="code1" target="code2">
