+++ docs/vertexnova/testbed/diagrams/architecture.drawio
@@ -4,25 +4,25 @@
             <root>
                 <mxCell id="0"/>
                 <mxCell id="1" parent="0"/>
-                <mxCell id="root" value="vnetemplate/&#10;(root)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#f5f5f5;strokeColor=#616161;strokeWidth=2;fontStyle=1;fontSize=12;fontColor=#1a1a1a;" parent="1" vertex="1">
+                <mxCell id="root" value="vnetestbed/&#10;(root)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#f5f5f5;strokeColor=#616161;strokeWidth=2;fontStyle=1;fontSize=12;fontColor=#1a1a1a;" parent="1" vertex="1">
                     <mxGeometry x="60" y="30" width="120" height="50" as="geometry"/>
                 </mxCell>
-                <mxCell id="include" value="include/&#10;vertexnova/template/&#10;(public API)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#e3f2fd;strokeColor=#1565c0;fontSize=11;align=left;spacingLeft=8;fontColor=#1a1a1a;" parent="1" vertex="1">
+                <mxCell id="include" value="include/&#10;vertexnova/testbed/&#10;(public API)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#e3f2fd;strokeColor=#1565c0;fontSize=11;align=left;spacingLeft=8;fontColor=#1a1a1a;" parent="1" vertex="1">
                     <mxGeometry x="40" y="120" width="160" height="50" as="geometry"/>
                 </mxCell>
-                <mxCell id="src" value="src/&#10;vertexnova/template/&#10;(implementation)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#e8f5e9;strokeColor=#2e7d32;fontSize=11;align=left;spacingLeft=8;fontColor=#1a1a1a;" parent="1" vertex="1">
+                <mxCell id="src" value="src/&#10;vertexnova/testbed/&#10;(implementation)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#e8f5e9;strokeColor=#2e7d32;fontSize=11;align=left;spacingLeft=8;fontColor=#1a1a1a;" parent="1" vertex="1">
                     <mxGeometry x="40" y="190" width="160" height="50" as="geometry"/>
                 </mxCell>
                 <mxCell id="tests" value="tests/&#10;(Google Test)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#fff3e0;strokeColor=#e65100;fontSize=11;fontColor=#1a1a1a;" parent="1" vertex="1">
                     <mxGeometry x="60" y="280" width="120" height="40" as="geometry"/>
                 </mxCell>
-                <mxCell id="examples" value="examples/&#10;(01_hello_template)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#fce4ec;strokeColor=#c2185b;fontSize=11;fontColor=#1a1a1a;" parent="1" vertex="1">
+                <mxCell id="examples" value="examples/&#10;(01_hello_testbed)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#fce4ec;strokeColor=#c2185b;fontSize=11;fontColor=#1a1a1a;" parent="1" vertex="1">
                     <mxGeometry x="40" y="360" width="160" height="40" as="geometry"/>
                 </mxCell>
                 <mxCell id="cmake" value="cmake/vnecmake/&#10;(submodule)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#f5f5f5;strokeColor=#616161;fontSize=11;fontColor=#1a1a1a;" parent="1" vertex="1">
                     <mxGeometry x="290" y="120" width="140" height="40" as="geometry"/>
                 </mxCell>
-                <mxCell id="build" value="CMake configure + build&#10;→ libvnetemplate.a&#10;→ tests, examples" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#e8f5e9;strokeColor=#2e7d32;strokeWidth=2;fontSize=12;fontStyle=1;fontColor=#1a1a1a;" parent="1" vertex="1">
+                <mxCell id="build" value="CMake configure + build&#10;→ libvnetestbed.a&#10;→ tests, examples" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#e8f5e9;strokeColor=#2e7d32;strokeWidth=2;fontSize=12;fontStyle=1;fontColor=#1a1a1a;" parent="1" vertex="1">
                     <mxGeometry x="560" y="180" width="220" height="70" as="geometry"/>
                 </mxCell>
                 <mxCell id="e1" style="endArrow=blockThin;endFill=1;html=1;strokeWidth=1;strokeColor=#616161;exitX=0.5;exitY=1;exitDx=0;exitDy=0;entryX=0.5;entryY=0;entryDx=0;entryDy=0;" parent="1" source="root" target="include" edge="1">
