+++ docs/vertexnova/testbed/diagrams/context.drawio
@@ -7,13 +7,13 @@
                 <mxCell id="app" value="C++ Application&#10;(Developer / User Code)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#ffffff;strokeColor=#1565c0;strokeWidth=2;fontStyle=1;fontSize=12;fontColor=#1a1a1a;" parent="1" vertex="1">
                     <mxGeometry x="20" y="320" width="200" height="80" as="geometry"/>
                 </mxCell>
-                <mxCell id="vne" value="VneTemplate&#10;Template Library&#10;(get_version, hello)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#ffffff;strokeColor=#e65100;strokeWidth=2;fontStyle=1;fontSize=14;fontColor=#1a1a1a;" parent="1" vertex="1">
+                <mxCell id="vne" value="VneTestbed&#10;Testbed Library&#10;(get_version, hello)" style="rounded=1;whiteSpace=wrap;html=1;fillColor=#ffffff;strokeColor=#e65100;strokeWidth=2;fontStyle=1;fontSize=14;fontColor=#1a1a1a;" parent="1" vertex="1">
                     <mxGeometry x="440" y="310" width="200" height="100" as="geometry"/>
                 </mxCell>
                 <mxCell id="r1" style="endArrow=blockThin;endFill=1;html=1;strokeWidth=2;strokeColor=#1565c0;exitX=1;exitY=0.5;exitDx=0;exitDy=0;entryX=0;entryY=0.5;entryDx=0;entryDy=0;" parent="1" source="app" target="vne" edge="1">
                     <mxGeometry relative="1" as="geometry"/>
                 </mxCell>
-                <mxCell id="r1label" value="uses&#10;vne::template_ns::get_version(), hello()" style="edgeLabel;html=1;align=center;verticalAlign=middle;resizable=0;points=[];fontSize=10;fontColor=#1a1a1a;backgroundColor=#ffffff;" parent="r1" connectable="0" vertex="1">
+                <mxCell id="r1label" value="uses&#10;vne::testbed_ns::get_version(), hello()" style="edgeLabel;html=1;align=center;verticalAlign=middle;resizable=0;points=[];fontSize=10;fontColor=#1a1a1a;backgroundColor=#ffffff;" parent="r1" connectable="0" vertex="1">
                     <mxGeometry x="-0.2" y="1" relative="1" as="geometry">
                         <mxPoint as="offset"/>
                     </mxGeometry>
