+++ docs/vertexnova/testbed/testbed.md
@@ -1,0 +1,71 @@
+# VertexNova Testbed
+
+## Overview
+
+VneTestbed is a minimal C++ project testbed for the VertexNova ecosystem. It provides a standard layout (include, src, tests, examples), CMake setup with vnecmake, optional internal deps (vnecommon, vnelogging), and a tiny API so you can build and run out of the box. Use it as a starting point for new libraries or applications.
+
+![System Context](diagrams/context.png)
+
+**Figure 1: Context Diagram**
+
+| Element | Description |
+|---------|-------------|
+| C++ Application | Developer/user code (tests, examples, or your app) that uses the testbed API |
+| VneTestbed | Testbed library; provides `get_version()`, `hello()`, and the project scaffold |
+
+## Project layout and build
+
+The testbed follows a standard directory layout and builds a static library, tests, and optional examples:
+
+![Project layout](diagrams/architecture.png)
+
+**Figure 2: Project layout and build**
+
+| Element | Description |
+|---------|-------------|
+| include/vertexnova/testbed/ | Public API headers (e.g. `testbed.h`) |
+| src/vertexnova/testbed/ | Implementation |
+| tests/ | Unit tests (Google Test) |
+| examples/ | Example apps (e.g. `01_hello_testbed`) |
+| cmake/vnecmake/ | CMake modules submodule |
+| deps/internal/, deps/external/ | Internal (vnecommon, vnelogging) and external (googletest) deps |
+| CMake configure + build | Produces `libvnetestbed.a`, tests, and examples |
+
+See the root [README.md](../../../README.md) for prerequisites, dependencies, and build commands.
+
+## API usage
+
+The public API lives in namespace `vne::testbed_ns` and exposes two functions:
+
+![API flow](diagrams/api.png)
+
+**Figure 3: API usage**
+
+| Step | Function | Description |
+|------|----------|-------------|
+| 1 | `get_version()` | Returns the project version string (e.g. from `VERSION` file). |
+| 2 | `hello()` | Returns a greeting string (minimal placeholder). |
+
+Example:
+
+```cpp
+#include <vertexnova/testbed/testbed.h>
+
+const char* ver = vne::testbed_ns::get_version();  // e.g. "1.0.0"
+const char* msg = vne::testbed_ns::hello();        // e.g. "Hello from VneTestbed"
+```
+
+## CMake options
+
+| Option | Default | Description |
+|--------|---------|-------------|
+| `VNE_TESTBED_TESTS` | ON | Build unit tests. |
+| `VNE_TESTBED_EXAMPLES` | ON (dev/top-level) / OFF (submodule) | Build examples (on by default in dev builds, off when used as a submodule). |
+| `WARNINGS_AS_ERRORS` | OFF | Treat compiler warnings as errors. |
+| `ENABLE_DOXYGEN` | OFF | Generate Doxygen documentation. |
+
+## Documentation
+
+- **This document:** `docs/vertexnova/testbed/testbed.md`
+- **Diagrams:** `docs/vertexnova/testbed/diagrams/` (Draw.io sources; export to PNG as described in [diagrams/README.md](diagrams/README.md))
+- **API reference:** Generated by Doxygen when `-DENABLE_DOXYGEN=ON` (see root README)
