+++ examples/CMakeLists.txt
@@ -3,9 +3,9 @@
 # Licensed under the Apache License, Version 2.0 (the "License")
 #==============================================================================
 
-if(NOT TARGET vnetemplate)
+if(NOT TARGET vnetestbed)
     return()
 endif()
 
-# 01 - Hello Template: minimal usage of vne::template_ns (hello, get_version)
-add_subdirectory(01_hello_template)
+# 01 - Hello Testbed: minimal usage of vne::testbed_ns (hello, get_version)
+add_subdirectory(01_hello_testbed)
